Skip saving and log read-only PDO parameters.
For PDO communication or mapping parameters which are read-only on a
node, the PdoMap.save() method currently aborts with an SDO exception.
However, individual parameters may simply not support configuration
changes.  Rely on the Object Dictionary to tell which ones are
expected to fail and skip them.  For communication record entries, log
what is not being saved.