Abort SDO server upload when the object has zero length.
A data size of zero bytes cannot be encoded in the "number of bytes
that do not contain data" semantics, as that would require n=4, which
is however limited to two bits (maximum value 3).  So the SDO
expedited upload protocol simply cannot convey that condition.  The
current implementation however still tries and thereby corrupts the
server command specifier field value.

Avoid that protocol violation by responding with an SDO abort code of
0800 0024h, No data available.

Add a test case for this condition, verifying that the client raises
the appropriate SdoAbortedError exception.