Avoid duplicate SDO abort messages on timeout.
For each call to SdoClient.read_response(), the timeout abort message
can either be sent immediately before raising the exception, or it is
sent later, after the exception has been handled e.g. by a retry.  Add
a boolean parameter to the function, which is usually False to skip
the immediate transmission of an SDO abort.  Only two cases actually
need it, passing the option as True.