Fix returned numbers in PDO.__iter__().
The base iterators of rx and tx return a 1-based sequence, while the
wrapping PDO class allows only mapping or communication record index
numbers to identify a PdoMap object.  Return the mapping record's
index from the iterator to avoid ambiguity.

Add tests to verify the returned objects are actually the RPDO and
TPDO members (in this order), with strictly increasing index numbers.